harness: absorb ensubstrate into mcp
